Convert more components to the new loading format.
