+++ app/views/games/index.html.erb
@@ -29,8 +29,8 @@
     </div>
   </div>
 
-  <div id="games-filter" class="is-fullwidth-mobile">
-    <games-filter></games-filter>
+  <div class="is-fullwidth-mobile">
+    <div data-vue-component="games-filter"></div>
   </div>
 </div>
