+++ app/views/games/new.html.erb
@@ -2,24 +2,15 @@
 
 <h1 class="title">Create a new game</h1>
 
-<div id='game-form'>
-  <% if @steam_app_id %>
-    <game-form
-      :submit-path="<%= games_url.to_json %>"
-      :rails-direct-uploads-path="<%= rails_direct_uploads_url.to_json %>"
-      :cancel-path="<%= games_url.to_json %>"
-      :create="true"
-      :name="<%= @name.to_json %>"
-      :steam_app_id="<%= @steam_app_id.to_i %>"
-    ></game-form>
-  <% else %>
-    <game-form
-      :submit-path="<%= games_url.to_json %>"
-      :rails-direct-uploads-path="<%= rails_direct_uploads_url.to_json %>"
-      :cancel-path="<%= games_url.to_json %>"
-      :create="true"
-      :name="<%= @name.to_json %>"
-    ></game-form>
-  <% end %>
-</div>
-
+<div data-vue-component="game-form"
+  data-vue-props="<%=
+    {
+      submitPath: games_url,
+      railsDirectUploadsPath: rails_direct_uploads_url,
+      successPath: games_url,
+      cancelPath: games_url,
+      create: true,
+      name: (@name ? @name : ''),
+      steamAppId: (@steam_app_id ? @steam_app_id : nil)
+    }.to_json
+  %>"></div>
